requirements.txt
pyyaml
paho-mqtt